requirements.txt
discord.py>=1.7.3
aiohttp>=3.8.1
PyNaCl>=1.4.0
